\name{compSampleAcivity}
\alias{compSampleAcivity}
\title{Compute the sample activity}
\usage{
  compSampleAcivity(str.ana, str.wd, dat.dir)
}
\arguments{
  \item{str.ana}{The name of the analysis \code{str.ana}}

  \item{str.wd}{Working directory for the analysis
  \code{str.wd}}

  \item{dat.dir}{Path to a data directory \code{dat.dir}}
}
\value{
  none
}
\description{
  A helper function to compute the sample activity prepare
  pdfs and tex files for the report
}
\examples{
library(rWipeTest)
}
\keyword{keywords}

