% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ComputeActivityAm241.R
\name{ComputeActivityAm241}
\alias{ComputeActivityAm241}
\title{Compute the activity for our Am241 standard}
\usage{
ComputeActivityAm241(dt)
}
\arguments{
\item{dt}{The analysis date in Y-M-D format \code{dt}}
}
\value{
nAct The current activity in dpm
}
\description{
Compute the current activity for the Am241 standard
certified on 2000-02-01 on the analysis date
}
\examples{
library(rWipeTest)
act <- ComputeActivityAm241("2013-11-01")
}
\keyword{keywords}

