% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/effPanelPplot.R
\name{eff.panel.plot}
\alias{eff.panel.plot}
\title{Create a panel plot of detector efficiency}
\usage{
eff.panel.plot(alpha, beta)
}
\arguments{
\item{alpha}{The input data frame for alpha standards \code{alpha}}

\item{beta}{The input data frame for alpha standards \code{beta}}
}
\value{
df.det.eff A data frame of detector efficiency
}
\description{
A helper function to compute a panel plot of detector
efficiency for alpha and beta irradiation.
}
\examples{
library(rWipeTest)
}
\keyword{keywords}

