% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sortAna.R
\name{sortAna}
\alias{sortAna}
\title{Sort the data and store the measurements in a data frame}
\usage{
sortAna(str.ana, str.wd, dat.dir)
}
\arguments{
\item{str.ana}{The name of the analysis \code{str.ana}}

\item{str.wd}{Working directory for the analysis \code{str.wd}}

\item{dat.dir}{Path to a data directory \code{dat.dir}}
}
\value{
none
}
\description{
A helper function to compute a panel plot of detector
efficiency for alpha and beta irradiation.
}
\examples{
library(rWipeTest)
}
\keyword{keywords}

