% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stderr.R
\name{stderr}
\alias{stderr}
\title{Compute the standard error of a vector}
\usage{
stderr(x)
}
\arguments{
\item{x}{The vector \code{x}}
}
\value{
stdError
}
\description{
A helper function to compute the standard error
}
\examples{
library(rWipeTest)
}
\keyword{keywords}

